Add company creation form
