+++ app/dashboard/company/page.tsx
@@ -1,11 +1,14 @@
+import CompanyCreationNavigatorButton from "@/components/company/CompanyCreationNavigatorButton";
+
 export default function CompanyPage() {
   return (
     <div className="w-full p-4">
       <div className="w-full">
         <h1 className="text-2xl font-semibold tracking-tight">Companies</h1>
-        <p className="text-muted-foreground mt-1 text-sm">
-          All the created companies are listed below
-        </p>
+        <p className="text-muted-foreground mt-1 text-sm">All the created companies are listed below</p>
+      </div>
+      <div className="flex w-full justify-end">
+        <CompanyCreationNavigatorButton />
       </div>
     </div>
   );
